Post edit state and update

+++ src/app.js
@@ -13,6 +13,10 @@ document.querySelector('.post-submit').addEventListener('click', addPost);
 // Listen for delete post
 document.querySelector('#posts').addEventListener('click', deletePost);
 
+// Listen for edit state
+document.querySelector('#posts').addEventListener('click', enableEdit);
+
+
 // Get Posts
 function getPosts(){
     http.get('http://localhost:3000/posts')
@@ -35,16 +39,37 @@ function addPost(){
     .catch(err => console.log(err));
 }
 
-//Delete Posts
+// Delete Posts
 function deletePost(e){
     e.preventDefault();
     if(e.target.parentElement.classList.contains('delete')){
     const id = e.target.parentElement.dataset.id;
-    http.delete(`http://localhost:3000/posts/${id}`)
+    if(confirm('Are you sure?')){
+        http.delete(`http://localhost:3000/posts/${id}`)
         .then(data => {
+            ui.showAlert('Post removed', 'alert alert-success');
             getPosts();
         })
         // .then(data => console.log(data))
         .catch(err => console.log(err));
         }
+    }
+}
+
+// Enable edit state
+function enableEdit(e){
+    if(e.target.parentElement.classList.contains('edit')){
+        const id = e.target.parentElement.dataset.id;
+        const body = e.target.parentElement.previousElementSibling.textContent;
+        const title = e.target.parentElement.previousElementSibling.previousElementSibling.textContent;
+        
+        const data = {
+            id,
+            title,
+            body
+        }
+
+        // Fill form with the current post
+        ui.fillForm(data);
+    }
 }
